add leading space
even though the dracut manual indicates that you can use multiple
filesystem definitions, they seem to be joined with string concatenation.
So, if you had filesystems+="ext4" in one file and filesystems+="ceph" in
another, these would be merged as "ext4ceph". This workaround seems to
work in both cases: when filesystems is set in another file and when it is
not set.